feat: implement investigate-potential-issue-in-stellar-token-minter-tests-for-smart-contract with tests and docs
- Add stellar_token_minter module (MAX_NFT_MINT_BATCH, NFT_MINT_FN_NAME, helpers).
- Wire withdraw() to mint_batch_is_full, bump_token_id_after_mint, and shared mint symbol.
- Re-export MAX_NFT_MINT_BATCH from lib.rs for existing imports.
- Consolidate orphaned tests into stellar_token_minter.test.rs and declare in lib.rs
  (withdraw_event_emission_test + stellar_token_minter_test never ran before).
- Replace stellar_token_minter.md with module-focused documentation; update
  withdraw_event_emission.md paths.
