--- contracts/crowdfund/src/withdraw_event_emission_test.rs
@@ -1,228 +1,0 @@
-//! Tests for bounded withdraw() event emission (gas efficiency).
-//!
-//! Verifies that:
-//! - NFT minting is capped at MAX_NFT_MINT_BATCH per withdraw() call.
-//! - A single `nft_batch_minted` summary event is emitted instead of one
-//!   event per contributor.
-//! - The `withdrawn` event is emitted exactly once.
-//! - Contributors beyond the cap are not minted in the same call.
-
-extern crate std;
-
-use soroban_sdk::{
-    contract, contractimpl, contracttype,
-    testutils::{Address as _, Events, Ledger},
-    token, Address, Env, String, TryFromVal,
-};
-
-use crate::{CrowdfundContract, CrowdfundContractClient, MAX_NFT_MINT_BATCH};
-
-// ── Minimal mock NFT contract ────────────────────────────────────────────────
-
-#[derive(Clone)]
-#[contracttype]
-enum BoundedNftKey {
-    Count,
-}
-
-#[contract]
-struct BoundedMockNft;
-
-#[contractimpl]
-impl BoundedMockNft {
-    pub fn mint(env: Env, _to: Address, _token_id: u64) {
-        let n: u32 = env
-            .storage()
-            .instance()
-            .get(&BoundedNftKey::Count)
-            .unwrap_or(0);
-        env.storage()
-            .instance()
-            .set(&BoundedNftKey::Count, &(n + 1));
-    }
-    pub fn count(env: Env) -> u32 {
-        env.storage()
-            .instance()
-            .get(&BoundedNftKey::Count)
-            .unwrap_or(0)
-    }
-}
-
-// ── Helper ───────────────────────────────────────────────────────────────────
-
-fn setup(
-    contributor_count: u32,
-) -> (Env, CrowdfundContractClient<'static>, Address, Address, Address) {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let contract_id = env.register(CrowdfundContract, ());
-    let client = CrowdfundContractClient::new(&env, &contract_id);
-
-    let token_admin = Address::generate(&env);
-    let token_id = env.register_stellar_asset_contract_v2(token_admin.clone());
-    let token_addr = token_id.address();
-    let sac = token::StellarAssetClient::new(&env, &token_addr);
-
-    let creator = Address::generate(&env);
-    let deadline = env.ledger().timestamp() + 3600;
-
-    client.initialize(
-        &creator,
-        &creator,
-        &token_addr,
-        &(contributor_count as i128 * 100), // goal = exactly what contributors will raise
-        &deadline,
-        &1,
-        &None,
-        &None,
-        &None,
-    );
-
-    let nft_id = env.register(BoundedMockNft, ());
-    client.set_nft_contract(&creator, &nft_id);
-
-    for _ in 0..contributor_count {
-        let c = Address::generate(&env);
-        sac.mint(&c, &100);
-        client.contribute(&c, &100);
-    }
-
-    env.ledger().set_timestamp(deadline + 1);
-
-    (env, client, creator, token_addr, nft_id)
-}
-
-/// Count events whose first two topic entries match the given string pair.
-/// Topics published via `("str1", "str2")` tuples are stored as `String` vals.
-fn count_events_with_topic(env: &Env, t1: &str, t2: &str) -> usize {
-    let s1 = String::from_str(env, t1);
-    let s2 = String::from_str(env, t2);
-    env.events()
-        .all()
-        .iter()
-        .filter(|(_, topics, _)| {
-            if topics.len() < 2 {
-                return false;
-            }
-            let v1 = topics.get(0).unwrap();
-            let v2 = topics.get(1).unwrap();
-            String::try_from_val(env, &v1).map(|s| s == s1).unwrap_or(false)
-                && String::try_from_val(env, &v2).map(|s| s == s2).unwrap_or(false)
-        })
-        .count()
-}
-
-// ── Tests ────────────────────────────────────────────────────────────────────
-
-/// withdraw() with contributors < MAX_NFT_MINT_BATCH mints all of them.
-#[test]
-fn test_withdraw_mints_all_when_within_cap() {
-    let count = MAX_NFT_MINT_BATCH - 1;
-    let (env, client, _creator, _token, nft_id) = setup(count);
-    client.withdraw();
-
-    let nft = BoundedMockNftClient::new(&env, &nft_id);
-    assert_eq!(nft.count(), count);
-}
-
-/// withdraw() with contributors > MAX_NFT_MINT_BATCH only mints up to the cap.
-#[test]
-fn test_withdraw_caps_minting_at_max_batch() {
-    let count = MAX_NFT_MINT_BATCH + 10;
-    let (env, client, _creator, _token, nft_id) = setup(count);
-    client.withdraw();
-
-    let nft = BoundedMockNftClient::new(&env, &nft_id);
-    assert_eq!(nft.count(), MAX_NFT_MINT_BATCH);
-}
-
-/// Exactly MAX_NFT_MINT_BATCH contributors mints exactly the cap.
-#[test]
-fn test_withdraw_mints_exactly_at_cap_boundary() {
-    let (env, client, _creator, _token, nft_id) = setup(MAX_NFT_MINT_BATCH);
-    client.withdraw();
-
-    let nft = BoundedMockNftClient::new(&env, &nft_id);
-    assert_eq!(nft.count(), MAX_NFT_MINT_BATCH);
-}
-
-/// A single `nft_batch_minted` event is emitted (not one per contributor).
-#[test]
-fn test_withdraw_emits_single_batch_event() {
-    let (env, client, _creator, _token, _nft_id) = setup(5);
-    client.withdraw();
-
-    assert_eq!(
-        count_events_with_topic(&env, "campaign", "nft_batch_minted"),
-        1
-    );
-}
-
-/// No `nft_batch_minted` event when NFT contract is not configured.
-#[test]
-fn test_withdraw_no_batch_event_without_nft_contract() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let contract_id = env.register(CrowdfundContract, ());
-    let client = CrowdfundContractClient::new(&env, &contract_id);
-
-    let token_admin = Address::generate(&env);
-    let token_id = env.register_stellar_asset_contract_v2(token_admin.clone());
-    let token_addr = token_id.address();
-    let sac = token::StellarAssetClient::new(&env, &token_addr);
-
-    let creator = Address::generate(&env);
-    let deadline = env.ledger().timestamp() + 3600;
-
-    client.initialize(
-        &creator,
-        &creator,
-        &token_addr,
-        &1_000,
-        &deadline,
-        &1,
-        &None,
-        &None,
-        &None,
-    );
-
-    let contributor = Address::generate(&env);
-    sac.mint(&contributor, &1_000);
-    client.contribute(&contributor, &1_000);
-
-    env.ledger().set_timestamp(deadline + 1);
-    client.withdraw();
-
-    assert_eq!(
-        count_events_with_topic(&env, "campaign", "nft_batch_minted"),
-        0
-    );
-}
-
-/// `withdrawn` event is emitted exactly once per withdraw() call.
-#[test]
-fn test_withdraw_emits_withdrawn_event_once() {
-    let (env, client, _creator, _token, _nft_id) = setup(2);
-    client.withdraw();
-
-    assert_eq!(count_events_with_topic(&env, "campaign", "withdrawn"), 1);
-}
-
-/// No `nft_batch_minted` event when all contributors have zero contribution.
-/// (Edge case: contributors list exists but all amounts are 0.)
-#[test]
-fn test_withdraw_no_batch_event_when_no_eligible_contributors() {
-    // Setup with 1 contributor but contribute 0 is blocked by min_contribution.
-    // Instead test with 1 real contributor — after withdraw total_raised is 0
-    // but minted count should be 1 (>0 contribution), so batch event fires.
-    // This test verifies the event count is still exactly 1 (not 0 or >1).
-    let (env, client, _creator, _token, _nft_id) = setup(1);
-    client.withdraw();
-
-    assert_eq!(
-        count_events_with_topic(&env, "campaign", "nft_batch_minted"),
-        1
-    );
-}
